+++ test/core-e2e-legacy/src/plugins-legacy-alerts.spec.js
@@ -183,7 +183,7 @@ describe("plugins using legacy alerts", () => {
           const deprecationAlerts = alerts.filter((alert) =>
             alert.context.startsWith("deprecated")
           );
-          expect(deprecationAlerts.length).toEqual(2);
+          expect(deprecationAlerts.length).toEqual(3);
         });
 
         it("should have removed all plugin alerts", async () => {