+++ test/core-e2e/src/scaffold.spec.js
@@ -16,6 +16,7 @@ const {
   waitForServer,
   fixturesFolder,
   removeConfigFile,
+  removeNewLines,
 } = require("./support/helpers");
 
 const configFile = path.resolve(fixturesFolder("temp"), "mocks.config.js");
@@ -53,9 +54,18 @@ describe("when nor config file nor mocks folder exists", () => {
         expect(fsExtra.existsSync(configFile)).toEqual(true);
       });
 
-      it("should have base as selected mock in config file", async () => {
-        const config = require(configFile);
+      it.skip("should have base as selected mock in config file", async () => {
+        const config = fsExtra.readFile(configFile, "utf-8");
         expect(config.mocks.selected).toEqual("base");
+      });
+
+      it("should have routes.delay as 0 in config file", async () => {
+        const config = await fsExtra.readFile(configFile, "utf-8");
+        expect(removeNewLines(config)).toEqual(
+          expect.stringContaining(
+            `routes: {    // Global delay to apply to routes    //delay: 0,  }`
+          )
+        );
       });
 
       it("should serve users under the /api/users path", async () => {
@@ -90,9 +100,18 @@ describe("when nor config file nor mocks folder exists", () => {
         await waitForServer();
       });
 
-      it("should have base as selected mock in config file", async () => {
+      it.skip("should have base as selected mock in config file", async () => {
         const config = require(configFile);
         expect(config.mocks.selected).toEqual("base");
+      });
+
+      it("should have routes.delay as 0 in config file", async () => {
+        const config = await fsExtra.readFile(configFile, "utf-8");
+        expect(removeNewLines(config)).toEqual(
+          expect.stringContaining(
+            `routes: {    // Global delay to apply to routes    //delay: 0,  }`
+          )
+        );
       });
 
       it("should serve users under the /api/users path", async () => {