+++ packages/config/CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 ### Removed
 
+## [Unreleased]
+- feat: Add `hasBeenSet` property to option
+
 ## [1.1.0] - 2022-06-27
 
 ### Added